==> 1 Partition Sets.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("ISIC_2019_Training_GroundTruth.csv", index_col=0)

df.sum()

stratifier = np.array(df)

from sklearn.model_selection import train_test_split
df_train, df_test = train_test_split(df.index, test_size=0.2, stratify=stratifier)

TEST = df.T[df_test].T

TRAIN = df.T[df_train].T

 TRAIN.shape, TEST.shape

TRAIN.to_csv("TRAIN.csv")
TEST.to_csv("TEST.csv")
