==> 2 Process Images.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Process-TEST-set" data-toc-modified-id="Process-TEST-set-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Process TEST set</a></span></li><li><span><a href="#Process-TRAIN-set" data-toc-modified-id="Process-TRAIN-set-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Process TRAIN set</a></span></li></ul></div>

Code can be made faster by changing order of "if" statements to reduce operations

import pandas as pd
import numpy as np
import cv2
from skimage import color
from skimage import io

from scipy import misc

TRAIN = pd.read_csv("TRAIN.csv", index_col=0)
TEST = pd.read_csv("TEST.csv", index_col=0)

# Grey world color correction algorithm
# from https://gist.github.com/shunsukeaihara/4603234

def grey_world(nimg):
    nimg = nimg.transpose(2, 0, 1).astype(np.uint32)
    mu_g = np.average(nimg[1])
    nimg[0] = np.minimum(nimg[0]*(mu_g/np.average(nimg[0])),255)
    nimg[2] = np.minimum(nimg[2]*(mu_g/np.average(nimg[2])),255)
    return nimg.transpose(1, 2, 0).astype(np.uint8)

## Process TEST set

import numpy as np

def process_image(name): # takes filename of the lesion sample
    img = io.imread("ISIC_2019_Training_Input/"+name+".jpg")
    img = grey_world(img)
    img = cv2.resize(img, dsize=(224, 224))#, interpolation=cv2.INTER_CUBIC)

    misc.imsave("Processed Images/TEST/"+ name+".jpg", img)

import time
from tqdm import tqdm_notebook
start = time.time()
for lesion in tqdm_notebook(TEST.index):
    process_image(lesion)
    
end = time.time()
print("Run time for "+ str(len(TEST.index)) + " images: " + str((end - start)/60)+" minutes")    

from glob import glob
filelist = glob("Processed Images/TEST/ISIC_*.jpg")
len(filelist)

## Process TRAIN set

def process_steps(name): # performs same steps as process_image but saves img in TRAIN
    img = io.imread("ISIC_2019_Training_Input/"+name+".jpg")
    img = grey_world(img)
    img = cv2.resize(img, dsize=(224, 224))#, interpolation=cv2.INTER_CUBIC)

    
    misc.imsave("Processed Images/TRAIN/"+ name+".jpg", img)

import time
from tqdm import tqdm_notebook

start = time.time()
for lesion in tqdm_notebook(TRAIN.index):
    process_steps(lesion)
     
end = time.time()
print("Run time for "+ str(len(TRAIN.index)) + " images: " + str(round((end - start)/60, 3))+" minutes")    


filelist = glob("Processed Images/TRAIN/ISIC_*.jpg")
len(filelist)
