==> 4 Prep Format.ipynb <==
this file prepares images for training on AWS GPUs

currently notebook configured to not use any data augmentation

import pandas as pd
import numpy as np
from glob import glob
from tqdm import tqdm_notebook
from skimage import io

TRAIN = pd.read_csv("TRAIN.csv", index_col=0)
TEST = pd.read_csv("TEST.csv", index_col=0)

AUGMENTED = pd.read_csv("AUGMENTED.csv", index_col=0)

TRAIN.index

# Prepare Training and Validation Sets

# numpy arrays for input and targets
imageList = []
targetList = []

imageValList = []
targetValList = []


# load kfold splits

foldList = np.load("kfold-splits.npy")

Fill in imageList and targetList with 4 folds

training_images = np.concatenate([foldList[0], foldList[1], foldList[2], foldList[3]])
for name in tqdm_notebook(training_images):

    imageList.append(io.imread("Processed Images/TRAIN/"+name+".jpg"))
    targetList.append(np.array(TRAIN.loc[name]))

    # Add augmented images also
    for i in range(4):
        imageList.append(io.imread("Processed Images/AUGMENTED/"+name+"_"+str(i)+".jpg"))
        targetList.append(np.array(AUGMENTED.loc[name+"_"+str(i)]))

Fill in imageValList and targetValList with remaining fold

validation_images = foldList[4]
for name in tqdm_notebook(validation_images):
    
    imageValList.append(io.imread("Processed Images/TRAIN/"+name+".jpg"))
    targetValList.append(np.array(TRAIN.loc[name]))

imageList = np.array(imageList)
targetList = np.array(targetList)

imageValList = np.array(imageValList)
targetValList = np.array(targetValList)

Compute weights for CNNs:

(weight of class) = (total number of samples)/(number of samples in class)

# training and validation images should have same proportions

np.sum(targetList, axis = 0)/np.sum(targetList)

# training and validation images should have same weights

np.repeat(np.sum(targetList), 9)/np.sum(targetList, axis=0)

#np.save("targetList", targetList)
#np.save("imageValList", imageValList)
#np.save("targetValList", targetValList)

np.mean(imageList, axis = tuple(range(imageList.ndim-1)))

np.mean(imageValList, axis = tuple(range(imageValList.ndim-1)))

# Prepare Test Set

testList = []
targetTestList = []

test_images = TEST.index
for name in tqdm_notebook(test_images):
    
    testList.append(io.imread("Processed Images/TEST/"+name+".jpg"))
    targetTestList.append(np.array(TEST.loc[name])) 

testList = np.array(testList)
targetTestList = np.array(targetTestList)

np.sum(targetTestList, axis = 0)/np.sum(targetTestList)

np.repeat(np.sum(targetTestList), 9)/np.sum(targetTestList, axis=0)

# Save all the arrays

np.savez_compressed("/Users/alecx/Downloads/AWS-LESIONDATA-2019",
                    imageList=imageList, targetList=targetList,
                    imageValList=imageValList, targetValList=targetValList,
                    testList=testList, targetTestList=targetTestList)


pd.DataFrame(targetList).sum()
